Adjust content of sidebar and pages related to that content  base on webapp,
