+++ reciapp/lib/pages/recipes_result_page.dart
@@ -93,11 +93,6 @@ class _RecipesResultState extends State<RecipesResult> {
                 ],
               ),
             ),
-            Container(
-                margin: EdgeInsets.symmetric(horizontal: 11),
-                child: Row(
-                    mainAxisAlignment: MainAxisAlignment.start,
-                    children: [FilterNewOldPopular()])),
             SizedBox(height: MediaQuery.of(context).size.height * 0.02),
             Container(
               margin: EdgeInsets.symmetric(horizontal: 3),
